requirements.txt
keras==2.9.0
numpy==1.24.2
Pillow==9.4.0
tensorflow==2.9.1